fix(web-api): pass unsigned bytes to ctype
api_v1_allmetrics.c and api_v1_alarms.c passed plain char values to isalnum() and toupper(). On signed-char platforms, high-bit request or metadata bytes promote to negative int values that are outside the C/POSIX ctype input domain unless they equal EOF.

This is a potential current bug: the endpoints are reachable today, and the source violates the ctype contract for non-ASCII bytes, but the worker input did not include a reproduced crash or corruption artifact.

Cast the byte arguments to uint8_t before calling isalnum() or toupper(), and keep the existing char assignment behavior after toupper(). This matches existing Netdata ctype-call patterns in nearby web and URL code.

Functional impact: ASCII behavior, endpoint dispatch, URL tokenization, response shape, buffer ownership, and locking are unchanged. High-bit bytes move from undefined behavior to defined locale-aware ctype behavior.

Performance impact: negligible. The change adds only integer casts on bytes already being processed.

Validation: git diff --cached --check passed. Isolated validation passed targeted ctype scans, low-priority CMake configure, and direct compiles of the changed translation units; broad Ninja object build was blocked by unrelated missing ACLK proto checkout content. Reviewer gate verified the ctype contract, all call paths, and absence of side effects.

+++ src/web/api/v1/api_v1_allmetrics.c
@@ -35,8 +35,8 @@ static inline size_t shell_name_copy(char *d, const char *s, size_t usable) {
     for(n = 0; *s && n < usable ; d++, s++, n++) {
         register char c = *s;
 
-        if(unlikely(!isalnum(c))) *d = '_';
-        else *d = (char)toupper(c);
+        if(unlikely(!isalnum((uint8_t)c))) *d = '_';
+        else *d = (char)toupper((uint8_t)c);
     }
     *d = '\0';
